phase 1 admin working on API security improvement.

+++ main/app/api/admin/winners/crud/remove/route.tsx
@@ -1,5 +1,6 @@
 import prisma from "@/app/libs/db";
 import { NextResponse } from "next/server";
+import jwt from "jsonwebtoken";
 
 interface Respo {
     success: boolean,
@@ -13,35 +14,80 @@ export async function DELETE(req: Request) {
         message: ''
     }
 
-    let sts: number = 400;
+    let sts: number = 200;
+    let isTrueAdminUser: boolean = false;
 
     try {
 
         const body = await req.json();
-        const { winner_type } = body;
+        const { token, winner_type } = body;
 
-        if (winner_type) {
-            const alreadyExist = await prisma.qF_Winners.findFirst({
-                where: {
-                    winner_type
-                }
-            });
-            if (alreadyExist !== null) {
-                await prisma.qF_Winners.delete({
+        if (token && winner_type) {
+
+            const res = jwt.verify(token as string, process.env.JWT_SECRET ?? "") as { is_admin_user: string };
+
+            if (res) {
+
+                const user_id = res.is_admin_user;
+
+                const fu__in__usrtblmdl = await prisma.qF_User.findFirst({
                     where: {
-                        winner_type
+                        AND: [
+                            {
+                                user_id
+                            },
+                            {
+                                role: "Admin"
+                            }
+                        ]
                     }
                 });
-                sts = 200;
-                resp = {
-                    success: true,
-                    message: "Winner Deleted Successfully!"
+                const fu__in__admntblmdl = await prisma.qF_Admin_User.findFirst({
+                    where: {
+                        admin_user_id: user_id,
+                    }
+                });
+
+                if (fu__in__usrtblmdl) {
+                    isTrueAdminUser = true;
+                } else {
+                    if (fu__in__admntblmdl) {
+                        isTrueAdminUser = true;
+                    } else {
+                        isTrueAdminUser = false;
+                    }
                 }
-            } else {
-                sts = 200;
-                resp = {
-                    success: false,
-                    message: "Winner Not Found!"
+
+                if (isTrueAdminUser) {
+                    const alreadyExist = await prisma.qF_Winners.findFirst({
+                        where: {
+                            winner_type
+                        }
+                    });
+                    if (alreadyExist !== null) {
+                        await prisma.qF_Winners.delete({
+                            where: {
+                                winner_type
+                            }
+                        });
+                        sts = 200;
+                        resp = {
+                            success: true,
+                            message: "Winner Deleted Successfully!"
+                        }
+                    } else {
+                        sts = 200;
+                        resp = {
+                            success: false,
+                            message: "Winner Not Found!"
+                        }
+                    }
+                } else {
+                    sts = 200;
+                    resp = {
+                        success: false,
+                        message: "User Not Found."
+                    }
                 }
             }
         } else {
@@ -55,10 +101,37 @@ export async function DELETE(req: Request) {
         return NextResponse.json(resp, { status: sts });
         //eslint-disable-next-line
     } catch (error: any) {
-        sts = 500;
-        resp = {
-            success: false,
-            message: error.message
+        // sts = 500;
+        // resp = {
+        //     success: false,
+        //     message: error.message
+        // }
+        if (error.message == "jwt expired") {
+            resp = {
+                success: false,
+                message: "Your session is expired, Please login again."
+            }
+        } else if (error.message == "jwt malformed" || error.message == "jwt must be a string") {
+            resp = {
+                success: false,
+                message: "Wrong information provided."
+            }
+        } else if (error.message == "invalid signature" || error.message == "invalid token") {
+            resp = {
+                success: false,
+                message: "Invalid information provided."
+            }
+        } else if (error.message == "jwt must be provided") {
+            sts = 400;
+            resp = {
+                success: false,
+                message: "Missing required fields."
+            }
+        } else {
+            resp = {
+                success: false,
+                message: error.message
+            }
         }
         return NextResponse.json(resp, { status: sts });
     }
